Fixed minor issues

+++ rational.cpp
@@ -16,20 +16,20 @@ long long gcd(long long x, long long y) // Euclidean algorithm
 
 void reduction(long long &x, long long &y)
 {
-    int nod = gcd(x, y);
-    if(!nod)
+    long long nod = gcd(x, y);
+    if (!nod)
     {
         y = 1;
         return;
     }
     x /= nod;
     y /= nod;
-    if(y < 0)
+    if (y < 0)
     {
         x *= -1;
         y *= -1;
     }
-    if(y==0)
+    if (y==0)
     {
         y = 1;
     }
@@ -45,7 +45,7 @@ Rational::Rational(long long n, long long d)
 {
     numerator = n;
     denominator = d;
-    if(!d)
+    if (!d)
         denominator = 1;
     reduction(numerator, denominator);
 }
@@ -61,7 +61,7 @@ void Rational::random(const long long &n)
     reduction(numerator, denominator);
 }
 
-std::ostream& operator<< (std::ostream &out, const Rational &r)
+std::ostream& operator << (std::ostream &out, const Rational &r)
 {
     out << r.numerator << '/' << r.denominator;
     return out;
@@ -103,7 +103,7 @@ Rational operator / (const Rational &r1, const Rational &r2)
     return r3;
 }
 
-bool operator ==(const Rational &r1, const Rational &r2)
+bool operator == (const Rational &r1, const Rational &r2)
 {
     if (r1.numerator * r2.denominator == r2.numerator * r1.denominator)
         return true;
